feat(order): enhance paidOrder method to accept discount and update order details fix(static): improve error handling for non-existent questions in deleteQuestion method

+++ src/static/static.service.ts
@@ -135,7 +135,7 @@ export class StaticService {
       where: { id },
     });
 
-    !ExistQuestion && new NotFoundException('Question not found');
+    if (!ExistQuestion) throw new NotFoundException('Question not found');
 
     const question = await this.prisma.questions.delete({ where: { id } });
     return new AppSuccess(question, 'Question deleted successfully');
